v26.5
fixed delete <-x_x->, modifi.. the design layout and added a feature if the user would like to add a stock
if they like the stock ticker info after searching the ticker symbol. Also made a pop-up to ensure double confirm if the
user would like to delete a stock or not.